v2.1.0

+++ Build-Dependencies.ps1
@@ -1,19 +1,25 @@
 param(
 	[Parameter(Mandatory = $true)]
 	[string]$ProjectPath,
+	[Parameter(Mandatory = $true)]
+	[string]$ModuleDirectory,
+	[Parameter(Mandatory = $false)]
+	[switch]$RemoveBaseLibrary,
 	[Parameter(Mandatory = $false)]
 	[switch]$IsExtensionModule
 )
 
 $projectRoot = Split-Path $ProjectPath -Parent
-$libFolder = "$projectRoot\lib"
+$libFolder = "$ModuleDirectory\lib"
 $projectName = (Get-Item $ProjectPath).BaseName
 
 # Builds all libraries that PoShLog depends on
 dotnet publish -c Release $ProjectPath -o $libFolder
 
 # Remove unecessary files
-Get-ChildItem $libFolder | Where-Object { $_.Name -like "*$projectName*" } | Remove-Item -Force
+if($RemoveBaseLibrary){
+	Get-ChildItem $libFolder | Where-Object { $_.Name -like "*$projectName*" } | Remove-Item -Force
+}
 
 if($IsExtensionModule){
 	Remove-Item -Path "$libFolder\Serilog.dll" -Force
